Update TestCreateOrg.java
Changed the TestCreateOrgButton into TestCreateOrg, since it was always on the way to making a organization.

Added the extra steps of changing to the create organization form, filling out the form and checking if we can submit the form. If we can submit the form, it means the test was successful. Since we can't get feedback from the backend yet (Since it's not made), we will use the button as active as our result if it worked or not.