Added a way to get organization project data
Added a way that you will be returned with the project you're asking for. As of this moment modules are not saved, so every project will appear to not have any modules when opened. This need to be added when submission of projects happen.

Due to how it's setup already the project are not gaining any information about the modules. Just the meta data about the project, we could add a way to gain information from an users project, the problem with that could be the users delete the project from their account.

Meaning you can't load in the modules, so the best way to implement it would add the modules when it get submitted to the organization.