Added hadOrgnization
After a project is submitted to an organization, the project will still have a member as their owner. This owner will see all the project that the organization has in their own personal projects as well.

The function will check if the project has a organization by checking if the organization_id column is not null. If it is null, the project gets returned.

Added in the service class that when a project get accepted by the organization the project owner will be transfered to the owner of the organization.